requirements.txt
requests==2.31.0
apify-client==1.6.1
openai==0.28.1
pyyaml==6.0.1
aiohttp==3.9.3
python-dotenv==1.0.1 